requirements.txt
python-telegram-bot==13.15
requests==2.32.3
flask==3.0.3
python-dotenv==1.0.1
psycopg2-binary==2.9.9